requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==42.0.7
defusedxml==0.7.1
distro==1.8.0
dj-database-url==2.1.0
Django==5.1.2
django-allauth==0.61.1
django-apscheduler==0.6.2
django-bootstrap5==23.3
django-cleanup==8.1.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-environ==0.11.2
django-extensions==3.2.3
django-htmx==1.17.3
django-ratelimit==4.1.0
djangorestframework==3.14.0
firebase-admin==6.1.0
frozenlist==1.4.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
install==1.3.5
jiter==0.5.0
markdown-it-py==2.2.0
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
Pillow==9.4.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.22.3
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.7.1
questionary==2.0.1
redis==5.2.0
reportlab==4.0.6
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.3.1
rsa==4.9
schedule==1.2.1
selenium==4.0.0b4
sendgrid==6.9.7
shell-gpt==0.7.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
stripe==5.5.0
tqdm==4.66.5
trio==0.25.0
trio-websocket==0.11.1
twilio==8.5.0
typer==0.7.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
urllib3-secure-extra==0.1.0
wcwidth==0.2.13
whitenoise==6.6.0
wsproto==1.2.0
yarl==1.9.2
